backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
qdrant-client>=1.7.0
openai>=1.10.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
httpx>=0.26.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
PyPDF2>=3.0.0
python-docx>=1.1.0
